small bug fixes and reworks to some stuff, saving with new enemy type and enemytype somewhat working

+++ src/Enemy.java
@@ -3,12 +3,12 @@ import java.util.ArrayList;
 abstract public class Enemy extends Moveable {//
 
     private double damage;
-    private int goldReward;
+    private double goldReward;
     private int respawnTimer;
     private int fullRespawnTimer;
 
 
-    Enemy(int x, int y, int width, int height, double health, double totalHealth, double damage, int goldReward, int respawnTimer, int fullRespawnTimer) {
+    Enemy(int x, int y, int width, int height, double health, double totalHealth, double damage, double goldReward, int respawnTimer, int fullRespawnTimer) {
         super(x, y, width, height, health, totalHealth);
         this.damage = damage;
         this.goldReward = goldReward;
@@ -17,7 +17,7 @@ abstract public class Enemy extends Moveable {//
     }
 
     // no respawn
-    Enemy(int x, int y, int width, int height, double health, double totalHealth, double damage, int goldReward) {
+    Enemy(int x, int y, int width, int height, double health, double totalHealth, double damage, double goldReward) {
         super(x, y, width, height, health, totalHealth);
         this.damage = damage;
         this.goldReward = goldReward;
@@ -283,7 +283,7 @@ abstract public class Enemy extends Moveable {//
         this.damage = damage;
     }
 
-    public int getGoldReward() {
+    public double getGoldReward() {
         return goldReward;
     }
